Separate EDF to data into a script and a module.
